Fixes failing CSVPropertyTableTest
There is no guarantee or reasonable assumption of order of files
returned from File#listFiles(). On my local instance the files are
returned in the reverse of the expected order but after digging through
some cursory documentation, I could not find an indication that the list
of files is sorted.

This patch refactors the test so as not to make an assumption of the
order of tests. This will fix failling tests where the order is not
preserved by File#listFiles().