Adds invokeWithHeaders step
This patch adds a new step that is based on the <invoke> but allows for
specifying custom HTTP headers to be sent with the requests. This
approach allows webtests to manage stored session cookies and pass
required information through the HTTP headers such as Cookies, CSRF
tokens, content type, etc.