Allows JSON objects as valid JSON
The VerifyJsonPath's method for determining if the result node is
validJson only checks if the node is an Array type, but will fail if the
node is an Object type. The method's docstring indicates that the
intended behavior should accept both arrays and objects, this patch
updates the method so as to accept objects as well.